backend/requirements.txt
asgiref
autopep8
Pillow
Django==2.1.*
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-heroku
django-allauth
djangorestframework
django-cors-headers
django-filter